+++ Cargo.toml
@@ -12,3 +12,4 @@ itertools = "0.14.0"
 criterion = "0.7.0"
 byteorder = "1.5.0"
 zstd = "0.13.3"
+once_cell = "1.21.3"
