+++ poulpy-backend/Cargo.toml
@@ -18,6 +18,7 @@ rand = {workspace = true}
 rand_distr = {workspace = true}
 rand_core = {workspace = true}
 byteorder = {workspace = true}
+once_cell = {workspace = true}
 rand_chacha = "0.9.0"
 
 [build-dependencies]
@@ -29,5 +30,5 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 
 [[bench]]
-name = "fft"
+name = "vmp"
 harness = false
