+++ poulpy-backend/src/cpu_spqlios/mod.rs
@@ -3,7 +3,7 @@ mod fft64;
 mod ntt120;
 
 #[cfg(test)]
-mod test;
+mod tests;
 
 pub use ffi::*;
 pub use fft64::*;
