--- poulpy-backend/src/cpu_spqlios/test/fft64/mod.rs
@@ -1,5 +1,0 @@
-pub mod svp;
-pub mod vec_znx;
-pub mod vec_znx_big;
-pub mod vec_znx_dft;
-pub mod vmp_pmat;
