+++ poulpy-hal/Cargo.toml
@@ -17,8 +17,10 @@ rand = {workspace = true}
 rand_distr = {workspace = true}
 rand_core = {workspace = true}
 byteorder = {workspace = true}
+once_cell = {workspace = true}
 rand_chacha = "0.9.0"
 bytemuck = "1.23.2"
+
 
 [build-dependencies]
 cmake = "0.1.54"
