+++ poulpy-hal/src/bench_suite/mod.rs
@@ -2,4 +2,4 @@ pub mod svp;
 pub mod vec_znx;
 pub mod vec_znx_big;
 pub mod vec_znx_dft;
-pub mod vmp_pmat;
+pub mod vmp;
