--- poulpy-hal/src/reference/fft64/svp/mod.rs
@@ -1,5 +1,0 @@
-mod svp_apply;
-mod svp_prepare;
-
-pub use svp_apply::*;
-pub use svp_prepare::*;
