--- poulpy-hal/src/reference/fft64/vmp/mod.rs
@@ -1,5 +1,0 @@
-mod vmp_apply;
-mod vmp_prepare;
-
-pub use vmp_apply::*;
-pub use vmp_prepare::*;
